pyproject.toml
# pyproject.toml

[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "augfeat"
version = "1.0.0"
description = "Data augmentation in features space"
readme = "README.md"
authors = [{ name = "Hubert", email = "hubert.hamelin@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["augmentation", "data", "machine-learning"]
dependencies = [
    'numpy',
    'scikit-learn',
    'tqdm',
    'matplotlib',
    'keras',
    'pandas',
    'Pillow',
    'idx2numpy'
]
requires-python = ">=3.9"

[project.urls]
Homepage = "https://github.com/HubertHamelin/augfeat"
